Move shared validation logic into collection/operations
