+++ lib/collection/src/operations/validation.rs
