+++ src/common/mod.rs
@@ -12,5 +12,3 @@ pub mod telemetry;
 pub mod telemetry_ops;
 #[allow(dead_code)] // May contain functions used in different binaries. Not actually dead
 pub mod telemetry_reporting;
-#[allow(dead_code)] // May contain functions used in different binaries. Not actually dead
-pub mod validation;
